Timers now only calculate nanoseconds.
Responsibility for the duration units lies with the reporters or backends.

+++ metrics-core/src/main/java/com/yammer/metrics/core/Histogram.java
@@ -131,33 +131,33 @@ public class Histogram implements Metric, Sampling, Summarizable {
      * @see com.yammer.metrics.core.Summarizable#max()
      */
     @Override
-    public double getMax() {
+    public long getMax() {
         if (getCount() > 0) {
             return max.get();
         }
-        return 0.0;
+        return 0;
     }
 
     /* (non-Javadoc)
      * @see com.yammer.metrics.core.Summarizable#min()
      */
     @Override
-    public double getMin() {
+    public long getMin() {
         if (getCount() > 0) {
             return min.get();
         }
-        return 0.0;
+        return 0;
     }
 
     /* (non-Javadoc)
      * @see com.yammer.metrics.core.Summarizable#mean()
      */
     @Override
-    public double getMean() {
+    public long getMean() {
         if (getCount() > 0) {
-            return sum.get() / (double) getCount();
+            return sum.get() / getCount();
         }
-        return 0.0;
+        return 0;
     }
 
     /* (non-Javadoc)
@@ -175,8 +175,8 @@ public class Histogram implements Metric, Sampling, Summarizable {
      * @see com.yammer.metrics.core.Summarizable#sum()
      */
     @Override
-    public double getSum() {
-        return (double) sum.get();
+    public long getSum() {
+        return sum.get();
     }
 
     @Override
