+++ metrics-core/src/main/java/com/yammer/metrics/stats/Snapshot.java
@@ -18,7 +18,7 @@ public class Snapshot {
     private static final double P99_Q = 0.99;
     private static final double P999_Q = 0.999;
 
-    private final double[] values;
+    private final long[] values;
 
     /**
      * Create a new {@link Snapshot} with the given values.
@@ -27,7 +27,7 @@ public class Snapshot {
      */
     public Snapshot(Collection<Long> values) {
         final Object[] copy = values.toArray();
-        this.values = new double[copy.length];
+        this.values = new long[copy.length];
         for (int i = 0; i < copy.length; i++) {
             this.values[i] = (Long) copy[i];
         }
@@ -39,8 +39,8 @@ public class Snapshot {
      *
      * @param values    an unordered set of values in the sample
      */
-    public Snapshot(double[] values) {
-        this.values = new double[values.length];
+    public Snapshot(long[] values) {
+        this.values = new long[values.length];
         System.arraycopy(values, 0, this.values, 0, values.length);
         Arrays.sort(this.values);
     }
@@ -51,13 +51,13 @@ public class Snapshot {
      * @param quantile    a given quantile, in {@code [0..1]}
      * @return the value in the distribution at {@code quantile}
      */
-    public double getValue(double quantile) {
+    public long getValue(double quantile) {
         if (quantile < 0.0 || quantile > 1.0) {
             throw new IllegalArgumentException(quantile + " is not in [0..1]");
         }
 
         if (values.length == 0) {
-            return 0.0;
+            return 0;
         }
 
         final double pos = quantile * (values.length + 1);
@@ -70,9 +70,9 @@ public class Snapshot {
             return values[values.length - 1];
         }
 
-        final double lower = values[(int) pos - 1];
-        final double upper = values[(int) pos];
-        return lower + (pos - floor(pos)) * (upper - lower);
+        final long lower = values[(int) pos - 1];
+        final long upper = values[(int) pos];
+        return (long) (lower + (pos - floor(pos)) * (upper - lower));
     }
 
     /**
@@ -89,7 +89,7 @@ public class Snapshot {
      *
      * @return the median value in the distribution
      */
-    public double getMedian() {
+    public long getMedian() {
         return getValue(MEDIAN_Q);
     }
 
@@ -98,7 +98,7 @@ public class Snapshot {
      *
      * @return the value at the 75th percentile in the distribution
      */
-    public double get75thPercentile() {
+    public long get75thPercentile() {
         return getValue(P75_Q);
     }
 
@@ -107,7 +107,7 @@ public class Snapshot {
      *
      * @return the value at the 95th percentile in the distribution
      */
-    public double get95thPercentile() {
+    public long get95thPercentile() {
         return getValue(P95_Q);
     }
 
@@ -116,7 +116,7 @@ public class Snapshot {
      *
      * @return the value at the 98th percentile in the distribution
      */
-    public double get98thPercentile() {
+    public long get98thPercentile() {
         return getValue(P98_Q);
     }
 
@@ -125,7 +125,7 @@ public class Snapshot {
      *
      * @return the value at the 99th percentile in the distribution
      */
-    public double get99thPercentile() {
+    public long get99thPercentile() {
         return getValue(P99_Q);
     }
 
@@ -134,7 +134,7 @@ public class Snapshot {
      *
      * @return the value at the 99.9th percentile in the distribution
      */
-    public double get999thPercentile() {
+    public long get999thPercentile() {
         return getValue(P999_Q);
     }
 
@@ -143,7 +143,7 @@ public class Snapshot {
      *
      * @return the entire set of values in the snapshot
      */
-    public double[] getValues() {
+    public long[] getValues() {
         return Arrays.copyOf(values, values.length);
     }
